+++ service/share/set.go
@@ -31,10 +31,10 @@ func (ps *limitedSet) Contains(p peer.ID) bool {
 	return ok
 }
 
-func (ps *limitedSet) Size() uint {
+func (ps *limitedSet) Size() int {
 	ps.lk.RLock()
 	defer ps.lk.RUnlock()
-	return uint(len(ps.ps))
+	return len(ps.ps)
 }
 
 // TryAdd attempts to add the given peer into the set.
